+++ json-circe/shared/src/test/scala-2/busymachines/pureharm/json/test/derivetest/semiAutoMelonJsonCodec.scala
@@ -20,25 +20,21 @@ import busymachines.pureharm.json._
 import busymachines.pureharm.json.implicits._
 import busymachines.pureharm.json.test._
 
-/** @author
-  *   Lorand Szakacs, https://github.com/lorandszakacs
-  * @since 11
-  *   Jun 2019
-  */
+@scala.annotation.nowarn
 private[test] object melonsDefaultSemiAutoDecoders {
   implicit val tasteDecoder:          Decoder[Taste]          = derive.enumerationDecoder[Taste]
   implicit val melonDecoder:          Decoder[Melon]          = derive.decoder[Melon]
   implicit val anarchistMelonDecoder: Decoder[AnarchistMelon] = derive.decoder[AnarchistMelon]
 }
 
-/** */
+@scala.annotation.nowarn
 private[test] object melonsDefaultSemiAutoEncoders {
   implicit val tasteEncoder:          Encoder[Taste]          = derive.enumerationEncoder[Taste]
   implicit val melonEncoder:          Encoder.AsObject[Melon] = derive.encoder[Melon]
   implicit val anarchistMelonEncoder: Encoder[AnarchistMelon] = derive.encoder[AnarchistMelon]
 }
 
-/** */
+@scala.annotation.nowarn
 private[test] object melonsDefaultSemiAutoCodecs {
   implicit val tasteCodec:          Codec[Taste]          = derive.enumerationCodec[Taste]
   implicit val melonCodec:          Codec[Melon]          = derive.codec[Melon]